refactor: extract prolly_btree_cursor_seek.c
Move table/index Moveto, custom-collation seek scans, serial-type and
cached seek-key helpers out of prolly_btree_cursor.c. Promote
payload-cache helpers and mergeLast for cross-TU use via
prolly_btree_int.h. Wire build, amalgamation, and lint.
