+++ src/prolly_btree_cursor.c
@@ -4,7 +4,7 @@
 
 /* Cursor navigation, seeking, comparison, and payload access. */
 
-static void cacheCurrentTreePayloadIfIntKey(BtCursor *pCur){
+void cacheCurrentTreePayloadIfIntKey(BtCursor *pCur){
   if( pCur->curIntKey ){
     const u8 *pVal; int nVal; int nAvail;
     CLEAR_CACHED_PAYLOAD(pCur);
@@ -17,7 +17,7 @@ static void cacheCurrentTreePayloadIfIntKey(BtCursor *pCur){
   }
 }
 
-static void cacheCurrentTreeStoredPayloadNonIntKey(BtCursor *pCur){
+void cacheCurrentTreeStoredPayloadNonIntKey(BtCursor *pCur){
   const u8 *pVal; int nVal;
   CLEAR_CACHED_PAYLOAD(pCur);
   cursorCurrentTreeValue(pCur, &pVal, &nVal);
@@ -28,27 +28,6 @@ static void cacheCurrentTreeStoredPayloadNonIntKey(BtCursor *pCur){
   }
 }
 
-
-static int keyInfoHasUnsupportedCollation(
-  const KeyInfo *pKeyInfo,
-  int nField
-){
-  int i;
-  if( !pKeyInfo ) return 0;
-  if( nField<=0 || nField>pKeyInfo->nAllField ){
-    nField = pKeyInfo->nAllField;
-  }
-  for(i=0; i<nField; i++){
-    const CollSeq *pColl = pKeyInfo->aColl[i];
-    if( !pColl || !pColl->zName ) continue;
-    if( sqlite3StrICmp(pColl->zName, "BINARY")!=0
-     && sqlite3StrICmp(pColl->zName, "NOCASE")!=0
-     && sqlite3StrICmp(pColl->zName, "RTRIM")!=0 ){
-      return 1;
-    }
-  }
-  return 0;
-}
 
 static int isDegenerateSchemaPayload(const u8 *pVal, int nVal){
   return nVal==6
@@ -415,7 +394,7 @@ static int mergeFirst(BtCursor *pCur, int *pRes){
   return mergeScan(pCur, 1, pRes);
 }
 
-static int mergeLast(BtCursor *pCur, int *pRes){
+int mergeLast(BtCursor *pCur, int *pRes){
   int rc = ensureCursorMutMapOrder(pCur);
   if( rc!=SQLITE_OK ) return rc;
   pCur->mmIdx = pCur->pMutMap->nEntries - 1;
@@ -722,987 +701,6 @@ int sqlite3BtreeIsEmpty(BtCursor *pCur, int *pRes){
   return pCur->pCurOps->xIsEmpty(pCur, pRes);
 }
 
-int prollyBtCursorTableMoveto(
-  BtCursor *pCur,
-  i64 intKey,
-  int bias,
-  int *pRes
-){
-  int rc;
-  struct TableEntry *pTE;
-  int rootIsEmpty;
-  int canUseAppendSeekFloor;
-  (void)bias;
-
-  assert( pCur->curIntKey );
-
-  pCur->nSeek++;
-  if( pCur->pBtree ) pCur->pBtree->nSeek++;
-  clearMergeCursorState(pCur);
-  CLEAR_CACHED_PAYLOAD(pCur);
-  CLEAR_CACHED_SEEK_KEY(pCur);
-  CLEAR_CACHED_COMPARE_KEY(pCur);
-
-  pTE = findTable(pCur->pBtree, pCur->pgnoRoot);
-  if( pTE && pTE->pPending && pCur->pMutMap!=(ProllyMutMap*)pTE->pPending ){
-    pCur->pMutMap = (ProllyMutMap*)pTE->pPending;
-  }
-  canUseAppendSeekFloor = pCur->pBtree
-    && pCur->pBtree->db
-    && !pCur->pBtree->db->autoCommit
-    && pCur->pBtree->db->pSavepoint==0;
-  rootIsEmpty = prollyHashIsEmpty(&pCur->pCur.root);
-  if( pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap) ){
-    ProllyMutMapEntry *pEntry = 0;
-    if( rootIsEmpty
-     && pCur->pMutMap->isIntKey
-     && pCur->pMutMap->appendSorted
-     && intKey > prollyMutMapEntryIntKey(
-                  &pCur->pMutMap->aEntries[pCur->pMutMap->nEntries-1])
-    ){
-      goto table_moveto_deferred_miss;
-    }
-    if( canUseAppendSeekFloor
-     && pTE
-     && pTE->appendSeekFloorValid
-     && prollyHashCompare(&pTE->root, &pTE->appendSeekRoot)==0
-     && pCur->pMutMap->isIntKey
-     && pCur->pMutMap->appendSorted
-     && intKey >= pTE->appendSeekFloor
-     && intKey > prollyMutMapEntryIntKey(
-                  &pCur->pMutMap->aEntries[pCur->pMutMap->nEntries-1])
-    ){
-      goto table_moveto_deferred_miss;
-    }
-    rc = prollyMutMapFindRc(pCur->pMutMap, 0, 0, intKey, &pEntry);
-    if( rc!=SQLITE_OK ) return rc;
-    if( pEntry && pEntry->op==PROLLY_EDIT_INSERT ){
-      if( pCur->isPinned ){
-        return SQLITE_CONSTRAINT_PINNED;
-      }
-      *pRes = 0;
-      setCursorToMutMapEntryPhys(pCur, (int)(pEntry - pCur->pMutMap->aEntries));
-      pCur->deferredTreeSeek = 1;
-      return SQLITE_OK;
-    }
-
-    /* intKey is delete-masked or absent from the mut map. Probe the tree
-    ** for a live exact hit first so equality lookups stay on the fast
-    ** path. */
-    if( !pEntry ){
-      refreshCursorRoot(pCur);
-      if( !prollyHashIsEmpty(&pCur->pCur.root) ){
-        int res = 0;
-        rc = prollyCursorSeekInt(&pCur->pCur, intKey, &res);
-        if( rc!=SQLITE_OK ) return rc;
-        if( res==0 ){
-          *pRes = 0;
-          pCur->eState = CURSOR_VALID;
-          pCur->curFlags |= BTCF_ValidNKey;
-          pCur->cachedIntKey = intKey;
-          CLEAR_CACHED_PAYLOAD(pCur);
-          cacheCurrentTreePayloadIfIntKey(pCur);
-          return SQLITE_OK;
-        }
-        if( res<0 && canUseAppendSeekFloor && pTE ){
-          if( !pTE->appendSeekFloorValid
-           || prollyHashCompare(&pTE->root, &pTE->appendSeekRoot)!=0
-           || intKey < pTE->appendSeekFloor
-          ){
-            pTE->appendSeekFloor = intKey;
-          }
-          pTE->appendSeekRoot = pTE->root;
-          pTE->appendSeekFloorValid = 1;
-        }
-      }
-    }
-
-    /* No live exact match. Report "not found" now but defer the merged
-    ** repositioning (which materializes the mut-map order) until something
-    ** actually consumes the cursor position: equality probes on write paths
-    ** never do, while range seeks consume it immediately via Eof/Next/Prev.
-    ** The raw tree position must not leak out of this path: it can sit on a
-    ** delete-masked row, skip mut-map-only rows, or see an empty root when
-    ** every live row is in the mut map.
-    **
-    ** The result must be -1, not +1: the VDBE treats res>0 as "the cursor
-    ** sits on a row above the key" and reads it with no Eof check, which a
-    ** deferred position cannot honor when nothing at or above the key
-    ** survives. With res<0 every consumer either steps forward through the
-    ** merge machinery or checks Eof before reading. */
-table_moveto_deferred_miss:
-    pCur->deferredMergedSeek = 1;
-    pCur->mmActive = 1;
-    pCur->cachedIntKey = intKey;
-    pCur->curFlags &= ~BTCF_ValidNKey;
-    pCur->eState = CURSOR_VALID;
-    *pRes = -1;
-    return SQLITE_OK;
-  }
-
-  if( rootIsEmpty ){
-    refreshCursorRoot(pCur);
-    rootIsEmpty = prollyHashIsEmpty(&pCur->pCur.root);
-  }
-  if( rootIsEmpty ){
-    /* Empty tables return res<0 with an invalid cursor, matching stock. */
-    *pRes = -1;
-    pCur->eState = CURSOR_INVALID;
-    return SQLITE_OK;
-  }
-
-  refreshCursorRoot(pCur);
-
-  rc = prollyCursorSeekInt(&pCur->pCur, intKey, pRes);
-  if( rc==SQLITE_OK ){
-    if( *pRes==0 ){
-      pCur->eState = CURSOR_VALID;
-      pCur->curFlags |= BTCF_ValidNKey;
-      pCur->cachedIntKey = intKey;
-      CLEAR_CACHED_PAYLOAD(pCur);
-      cacheCurrentTreePayloadIfIntKey(pCur);
-    } else if( pCur->pCur.eState==PROLLY_CURSOR_VALID ){
-      if( *pRes<0 && canUseAppendSeekFloor ){
-        if( pTE ){
-          if( !pTE->appendSeekFloorValid
-           || prollyHashCompare(&pTE->root, &pTE->appendSeekRoot)!=0
-           || intKey < pTE->appendSeekFloor
-          ){
-            pTE->appendSeekFloor = intKey;
-          }
-          pTE->appendSeekRoot = pTE->root;
-          pTE->appendSeekFloorValid = 1;
-        }
-      }
-      pCur->eState = CURSOR_VALID;
-      pCur->curFlags &= ~BTCF_ValidNKey;
-      cacheCurrentTreePayloadIfIntKey(pCur);
-    } else {
-      pCur->eState = CURSOR_INVALID;
-    }
-  }
-  return rc;
-}
-int sqlite3BtreeTableMoveto(
-  BtCursor *pCur,
-  i64 intKey,
-  int bias,
-  int *pRes
-){
-  if( !pCur ) return SQLITE_OK;
-  return pCur->pCurOps->xTableMoveto(pCur, intKey, bias, pRes);
-}
-
-u32 btreeSerialType(Mem *pMem, u32 *pLen){
-  int flags = pMem->flags;
-  if( flags & MEM_Null ){ *pLen = 0; return SERIAL_TYPE_NULL; }
-  if( flags & MEM_Int ){
-    i64 v = pMem->u.i;
-    if( v==0 ){ *pLen = 0; return SERIAL_TYPE_ZERO; }
-    if( v==1 ){ *pLen = 0; return SERIAL_TYPE_ONE; }
-    if( v>=-128 && v<=127 ){ *pLen = 1; return SERIAL_TYPE_INT8; }
-    if( v>=-32768 && v<=32767 ){ *pLen = 2; return SERIAL_TYPE_INT16; }
-    if( v>=-8388608 && v<=8388607 ){ *pLen = 3; return SERIAL_TYPE_INT24; }
-    if( v>=-2147483648LL && v<=2147483647LL ){ *pLen = 4; return SERIAL_TYPE_INT32; }
-    if( v>=-140737488355328LL && v<=140737488355327LL ){ *pLen = 6; return SERIAL_TYPE_INT48; }
-    *pLen = 8; return SERIAL_TYPE_INT64;
-  }
-  if( flags & MEM_Real ){ *pLen = 8; return SERIAL_TYPE_FLOAT64; }
-  if( flags & MEM_Str ){
-    u32 n = (u32)pMem->n;
-    *pLen = n;
-    return n*2 + SERIAL_TYPE_TEXT_BASE;
-  }
-  if( flags & MEM_Blob ){
-    u32 n = (u32)pMem->n;
-    *pLen = n;
-    return n*2 + SERIAL_TYPE_BLOB_BASE;
-  }
-  *pLen = 0; return SERIAL_TYPE_NULL;
-}
-
-static int findMatchingMutMapEntry(
-  ProllyMutMap *pMap,
-  KeyInfo *pKeyInfo,
-  UnpackedRecord *pIdxKey,
-  const u8 *pSortKey,
-  int nSortKey,
-  int bExactKey,
-  ProllyMutMapEntry **ppMatch,
-  int *pCmp
-){
-  int rc = SQLITE_OK;
-  int cmp = 0;
-  ProllyMutMapEntry *pMatch = 0;
-  u8 *pRecBuf = 0;
-  int nRecBufAlloc = 0;
-  int lo = 0, found = 0;
-
-  *ppMatch = 0;
-  *pCmp = 0;
-  if( !pMap || prollyMutMapIsEmpty(pMap) ){
-    return SQLITE_OK;
-  }
-
-  if( bExactKey
-   || (pKeyInfo && pIdxKey->nField >= pKeyInfo->nAllField) ){
-    ProllyMutMapEntry *pEntry = 0;
-    rc = prollyMutMapFindRc(pMap, pSortKey, nSortKey, 0, &pEntry);
-    if( rc!=SQLITE_OK ) return rc;
-    if( pEntry && pEntry->op==PROLLY_EDIT_INSERT ){
-      *ppMatch = pEntry;
-    }
-    return SQLITE_OK;
-  }
-
-  rc = prollyMutMapResolveSortedPos(pMap, pSortKey, nSortKey, 0,
-                                    &lo, &found);
-
-  while( rc==SQLITE_OK && lo < pMap->nEntries ){
-    ProllyMutMapEntry *pEntry = prollyMutMapEntryAt(pMap, lo);
-    const u8 *pRec = pEntry->pVal;
-    int nRec = pEntry->nVal;
-    int cmpLen;
-    int prefixCmp;
-
-    if( pMap->isIntKey ){
-      lo++;
-      continue;
-    }
-    cmpLen = pEntry->nKey < nSortKey ? pEntry->nKey : nSortKey;
-    prefixCmp = memcmp(pEntry->pKey, pSortKey, cmpLen);
-    if( prefixCmp>0 && pIdxKey->default_rc<0 ){
-      if( pEntry->op==PROLLY_EDIT_INSERT ){
-        pMatch = pEntry;
-        cmp = 1;
-      }else{
-        lo++;
-        continue;
-      }
-      break;
-    }
-    if( prefixCmp>0 ){
-      break;
-    }
-    if( prefixCmp==0 && pEntry->nKey < nSortKey ){
-      break;
-    }
-    if( nRec==0 ){
-      rc = recordFromSortKeyBufferColl(pEntry->pKey, pEntry->nKey,
-                                    pKeyInfo,
-                                    &pRecBuf, &nRecBufAlloc, &nRec);
-      if( rc!=SQLITE_OK ) break;
-      pRec = pRecBuf;
-    }
-    pIdxKey->eqSeen = 0;
-    cmp = sqlite3VdbeRecordCompare(nRec, pRec, pIdxKey);
-    if( cmp!=0 && !pIdxKey->eqSeen ){
-      break;
-    }
-    if( pEntry->op==PROLLY_EDIT_INSERT ){
-      pMatch = pEntry;
-      break;
-    }
-    lo++;
-  }
-
-  sqlite3_free(pRecBuf);
-  if( rc==SQLITE_OK && pMatch ){
-    *ppMatch = pMatch;
-    *pCmp = cmp;
-  }
-  return rc;
-}
-
-static int scanMutMapForCustomCollation(
-  BtCursor *pCur,
-  ProllyMutMap *pMap,
-  UnpackedRecord *pIdxKey,
-  ProllyMutMapEntry **ppMatch,
-  int *pCmp
-){
-  int i;
-  int rc = SQLITE_OK;
-  u8 *pRecBuf = 0;
-  int nRecBufAlloc = 0;
-
-  *ppMatch = 0;
-  *pCmp = 0;
-  if( !pMap || prollyMutMapIsEmpty(pMap) || pMap->isIntKey ){
-    return SQLITE_OK;
-  }
-
-  for(i=0; i<pMap->nEntries; i++){
-    ProllyMutMapEntry *pEntry = &pMap->aEntries[i];
-    const u8 *pRec = pEntry->pVal;
-    int nRec = pEntry->nVal;
-    int cmp;
-
-    if( pEntry->op!=PROLLY_EDIT_INSERT ){
-      continue;
-    }
-    if( nRec==0 ){
-      rc = recordFromSortKeyBufferColl(pEntry->pKey, pEntry->nKey,
-                                    pCur->pKeyInfo,
-                                    &pRecBuf, &nRecBufAlloc, &nRec);
-      if( rc!=SQLITE_OK ) break;
-      pRec = pRecBuf;
-    }
-    pIdxKey->eqSeen = 0;
-    cmp = sqlite3VdbeRecordCompare(nRec, pRec, pIdxKey);
-    if( cmp==0 || pIdxKey->eqSeen ){
-      *ppMatch = pEntry;
-      *pCmp = cmp;
-      break;
-    }
-  }
-
-  sqlite3_free(pRecBuf);
-  return rc;
-}
-
-static int scanTreeForCustomCollation(
-  BtCursor *pCur,
-  UnpackedRecord *pIdxKey,
-  int *pFound,
-  int *pCmp
-){
-  int rc;
-  int res = 0;
-  u8 *pRecBuf = pCur->pMovetoRec;
-  int nRecBufAlloc = pCur->nMovetoRecAlloc;
-
-  *pFound = 0;
-  *pCmp = 0;
-  rc = prollyCursorFirst(&pCur->pCur, &res);
-  while( rc==SQLITE_OK && res==0 && prollyCursorIsValid(&pCur->pCur) ){
-    const u8 *pKey = 0;
-    int nKey = 0;
-    const u8 *pRec = 0;
-    int nRec = 0;
-    int cmp;
-    int isDeleted = 0;
-
-    rc = prollyCursorCheckInterrupt(pCur);
-    if( rc!=SQLITE_OK ) break;
-    prollyCursorKey(&pCur->pCur, &pKey, &nKey);
-    if( pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap) ){
-      ProllyMutMapEntry *pEntry = 0;
-      rc = prollyMutMapFindRc(pCur->pMutMap, pKey, nKey, 0, &pEntry);
-      if( rc!=SQLITE_OK ) break;
-      if( pEntry && pEntry->op==PROLLY_EDIT_DELETE ) isDeleted = 1;
-    }
-    if( !isDeleted ){
-      prollyCursorValue(&pCur->pCur, &pRec, &nRec);
-      if( nRec==0 ){
-        rc = recordFromSortKeyBufferColl(pKey, nKey, pCur->pKeyInfo,
-                                         &pRecBuf, &nRecBufAlloc, &nRec);
-        if( rc!=SQLITE_OK ) break;
-        pRec = pRecBuf;
-      }
-      pIdxKey->eqSeen = 0;
-      cmp = sqlite3VdbeRecordCompare(nRec, pRec, pIdxKey);
-      if( cmp==0 || pIdxKey->eqSeen ){
-        *pFound = 1;
-        *pCmp = cmp;
-        break;
-      }
-    }
-    rc = prollyCursorNext(&pCur->pCur);
-  }
-
-  pCur->pMovetoRec = pRecBuf;
-  pCur->nMovetoRecAlloc = nRecBufAlloc;
-  return rc;
-}
-
-int prollyBtCursorIndexMoveto(
-  BtCursor *pCur,
-  UnpackedRecord *pIdxKey,
-  int *pRes
-){
-  int rc;
-
-  assert( !pCur->curIntKey );
-
-  if( pCur->pBtree ) pCur->pBtree->nSeek++;
-
-  clearMergeCursorState(pCur);
-  CLEAR_CACHED_PAYLOAD(pCur);
-  CLEAR_CACHED_SEEK_KEY(pCur);
-  CLEAR_CACHED_COMPARE_KEY(pCur);
-
-  refreshCursorRoot(pCur);
-
-  {
-    int treeFound = 0, mutFound = 0;
-    int treeCmp = 0, mutCmp = 0;
-    const u8 *mutKey = 0;
-    int mutNKey = 0;
-    ProllyMutMapEntry *mutE = 0;
-    int mutFromCursorMap = 0;
-    int exactMutMapKey = 0;
-
-    u8 *pSerKey = 0;
-    int nSerKey = 0;
-    u8 *pSortKey = 0;
-    int nSortKey = 0;
-    int nSeekKeyField = 0;
-    if( pCur->pKeyInfo && pIdxKey->nField < pCur->pKeyInfo->nAllField ){
-      nSeekKeyField = (int)pIdxKey->nField;
-    }
-    /* Table-root range seeks must ignore probe fields beyond the PK. */
-    if( pCur->isTableRoot && pCur->pKeyInfo
-     && pIdxKey->default_rc != 0
-     && pIdxKey->nField > pCur->pKeyInfo->nKeyField
-     && (nSeekKeyField==0 || nSeekKeyField > pCur->pKeyInfo->nKeyField) ){
-      nSeekKeyField = pCur->pKeyInfo->nKeyField;
-    }
-    if( unpackedRecordCanUseIntSortKey(
-            pCur, pIdxKey,
-            nSeekKeyField>0 ? nSeekKeyField : (int)pIdxKey->nField) ){
-      rc = sortKeyFromUnpackedIntRecordBuffer(
-          pIdxKey, nSeekKeyField>0 ? nSeekKeyField : (int)pIdxKey->nField,
-          &pCur->pSeekSortKey, &pCur->nSeekSortKeyAlloc, &nSortKey);
-    }else{
-      rc = sortKeyFromMemPrefixCollBuffer(
-          pIdxKey->aMem, (int)pIdxKey->nField, nSeekKeyField,
-          pCur->pKeyInfo,
-          &pCur->pSeekSortKey, &pCur->nSeekSortKeyAlloc, &nSortKey);
-      if( rc==SQLITE_NOTFOUND ){
-        rc = serializeUnpackedRecordBuffer(
-            pIdxKey, &pCur->pSeekRecord, &pCur->nSeekRecordAlloc, &nSerKey);
-        if( rc!=SQLITE_OK ) return rc;
-        pSerKey = pCur->pSeekRecord;
-        rc = sortKeyFromRecordPrefixCollBuffer(
-            pSerKey, nSerKey, nSeekKeyField, pCur->pKeyInfo,
-            &pCur->pSeekSortKey, &pCur->nSeekSortKeyAlloc, &nSortKey);
-      }
-    }
-    if( rc!=SQLITE_OK ) return rc;
-    pSortKey = pCur->pSeekSortKey;
-    pCur->nSeekSortKey = nSortKey;
-    pCur->nSeekKeyField = nSeekKeyField;
-
-    if( pCur->pKeyInfo
-     && nSeekKeyField == pCur->pKeyInfo->nKeyField ){
-      if( pCur->isTableRoot ){
-        exactMutMapKey = 1;
-      }
-    }
-
-    if( pCur->pKeyInfo
-     && (exactMutMapKey || pIdxKey->nField >= pCur->pKeyInfo->nAllField)
-     && keyInfoHasUnsupportedCollation(
-          pCur->pKeyInfo,
-          nSeekKeyField>0 ? nSeekKeyField : (int)pIdxKey->nField) ){
-      struct TableEntry *pTE = findTable(pCur->pBtree, pCur->pgnoRoot);
-      ProllyMutMap *pPending = pTE ? (ProllyMutMap*)pTE->pPending : 0;
-      ProllyMutMapEntry *pEntry = 0;
-      int cmp = 0;
-      int treeFound = 0;
-
-      rc = scanMutMapForCustomCollation(
-          pCur, (ProllyMutMap*)pCur->pMutMap, pIdxKey, &pEntry, &cmp);
-      if( rc!=SQLITE_OK ) return rc;
-      if( pEntry ){
-        setCursorToMutMapEntryPhys(
-            pCur, (int)(pEntry - pCur->pMutMap->aEntries));
-        *pRes = cmp;
-        pIdxKey->eqSeen = 1;
-        return SQLITE_OK;
-      }
-
-      if( pPending && pPending!=pCur->pMutMap ){
-        rc = scanMutMapForCustomCollation(
-            pCur, pPending, pIdxKey, &pEntry, &cmp);
-        if( rc!=SQLITE_OK ) return rc;
-        if( pEntry ){
-          const u8 *pVal = pEntry->pVal;
-          int nVal = pEntry->nVal;
-          if( nVal==0 ){
-            rc = cacheCursorPayloadReconstructed(
-                pCur, pEntry->pKey, pEntry->nKey);
-          }else{
-            rc = cacheCursorPayloadCopy(pCur, pVal, nVal);
-          }
-          if( rc!=SQLITE_OK ) return rc;
-          pCur->eState = CURSOR_VALID;
-          *pRes = cmp;
-          pIdxKey->eqSeen = 1;
-          return SQLITE_OK;
-        }
-      }
-
-      rc = scanTreeForCustomCollation(pCur, pIdxKey, &treeFound, &cmp);
-      if( rc!=SQLITE_OK ) return rc;
-      if( treeFound ){
-        pCur->eState = CURSOR_VALID;
-        cacheCurrentTreeStoredPayloadNonIntKey(pCur);
-        *pRes = cmp;
-        pIdxKey->eqSeen = 1;
-        return SQLITE_OK;
-      }
-    }
-
-    if( pCur->pKeyInfo
-     && (exactMutMapKey || pIdxKey->nField >= pCur->pKeyInfo->nAllField) ){
-      struct TableEntry *pTE = findTable(pCur->pBtree, pCur->pgnoRoot);
-      ProllyMutMap *pPending = pTE ? (ProllyMutMap*)pTE->pPending : 0;
-      ProllyMutMapEntry *pEntry = 0;
-      if( pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap) ){
-        rc = prollyMutMapFindRc(pCur->pMutMap, pSortKey, nSortKey, 0, &pEntry);
-        if( rc!=SQLITE_OK ) return rc;
-        if( pEntry && pEntry->op==PROLLY_EDIT_INSERT ){
-          if( pCur->isPinned ){
-            return SQLITE_CONSTRAINT_PINNED;
-          }
-          setCursorToMutMapEntryPhys(
-              pCur, (int)(pEntry - pCur->pMutMap->aEntries));
-          /* The tree side is still wherever the last operation left it. A
-          ** later step must re-seek it to this key before merging, or it
-          ** feeds stale entries into the merge scan. */
-          pCur->deferredTreeSeek = 1;
-          *pRes = 0;
-          pIdxKey->eqSeen = 1;
-          return SQLITE_OK;
-        }
-      }
-      if( pPending && pPending!=pCur->pMutMap
-       && !prollyMutMapIsEmpty(pPending) ){
-        rc = prollyMutMapFindRc(pPending, pSortKey, nSortKey, 0, &pEntry);
-        if( rc!=SQLITE_OK ) return rc;
-        if( pEntry && pEntry->op==PROLLY_EDIT_INSERT ){
-          if( pCur->isPinned ){
-            return SQLITE_CONSTRAINT_PINNED;
-          }
-          pCur->pMutMap = pPending;
-          setCursorToMutMapEntryPhys(
-              pCur, (int)(pEntry - pPending->aEntries));
-          pCur->deferredTreeSeek = 1;
-          *pRes = 0;
-          pIdxKey->eqSeen = 1;
-          return SQLITE_OK;
-        }
-      }
-    }
-
-    {
-      int seekRes = 0;
-      rc = prollyCursorSeekBlob(&pCur->pCur, pSortKey, nSortKey, &seekRes);
-      if( rc==SQLITE_OK
-       && seekRes==0
-       && pCur->pCur.eState==PROLLY_CURSOR_VALID
-       && pCur->pKeyInfo
-       && (exactMutMapKey || pIdxKey->nField >= pCur->pKeyInfo->nAllField) ){
-        int isDeleted = 0;
-        if( pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap) ){
-          ProllyMutMapEntry *mmE = 0;
-          rc = prollyMutMapFindRc(pCur->pMutMap, pSortKey, nSortKey, 0, &mmE);
-          if( rc!=SQLITE_OK ) return rc;
-          if( mmE && mmE->op==PROLLY_EDIT_DELETE ) isDeleted = 1;
-        }
-        if( !isDeleted ){
-          *pRes = 0;
-          pIdxKey->eqSeen = 1;
-          pCur->eState = CURSOR_VALID;
-          cacheCurrentTreeStoredPayloadNonIntKey(pCur);
-          return SQLITE_OK;
-        }
-      }
-    }
-    /* A failed tree seek must not fall through to the mut-map merge below:
-    ** that path resets rc and reports "row not found" for what was really
-    ** an I/O or allocation failure. */
-    if( rc!=SQLITE_OK ) return rc;
-    if( pCur->pCur.eState==PROLLY_CURSOR_VALID ){
-      int iLevel = pCur->pCur.iLevel;
-      ProllyCacheEntry *pLeaf = pCur->pCur.aLevel[iLevel].pEntry;
-      int lo = pCur->pCur.aLevel[iLevel].idx;
-      int nItems = pLeaf->node.nItems;
-
-      int bestIdx = -1;
-      int bestCmp = 0;
-      {
-        u8 *pRecBuf = pCur->pMovetoRec;
-        int nRecBufAlloc = pCur->nMovetoRecAlloc;
-        int i;
-
-       for(;;){
-        for( i = lo; i < nItems; i++ ){
-          const u8 *pSK; int nSK;
-          const u8 *pVal; int nVal;
-          int recCmp;
-          prollyNodeKey(&pLeaf->node, i, &pSK, &nSK);
-
-          {
-            int cmpLen = nSK < nSortKey ? nSK : nSortKey;
-            int prefixCmp = memcmp(pSK, pSortKey, cmpLen);
-            if( prefixCmp > 0 ){
-              if( bestIdx < 0 ){
-                if( pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap) ){
-                  ProllyMutMapEntry *mmE = 0;
-                  rc = prollyMutMapFindRc(pCur->pMutMap, pSK, nSK, 0, &mmE);
-                  if( rc!=SQLITE_OK ) break;
-                  if( mmE && mmE->op==PROLLY_EDIT_DELETE ){
-                    /* Delete-masked row past the seek key: the next live row
-                    ** is still the positioning answer, so keep scanning. */
-                    continue;
-                  }
-                }
-                bestIdx = i;
-                if( nSeekKeyField>0 ){
-                  pIdxKey->eqSeen = 0;
-                  bestCmp = 1;
-                }else{
-                  const u8 *pVal2; int nVal2;
-                  prollyNodeValue(&pLeaf->node, i, &pVal2, &nVal2);
-                  if( nVal2==0 ){
-                    rc = recordFromSortKeyBufferColl(
-                        pSK, nSK, pCur->pKeyInfo,
-                        &pRecBuf, &nRecBufAlloc, &nVal2);
-                    if( rc!=SQLITE_OK ) break;
-                    pVal2 = pRecBuf;
-                  }
-                  pIdxKey->eqSeen = 0;
-                  bestCmp = sqlite3VdbeRecordCompare(nVal2, pVal2, pIdxKey);
-                }
-              }
-              break;
-            }
-            if( nSeekKeyField>0 && prefixCmp==0 && nSK>=nSortKey ){
-              if( pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap) ){
-                ProllyMutMapEntry *mmE = 0;
-                rc = prollyMutMapFindRc(pCur->pMutMap, pSK, nSK, 0, &mmE);
-                if( rc!=SQLITE_OK ) break;
-                if( mmE && mmE->op==PROLLY_EDIT_DELETE ){
-                  continue;
-                }
-              }
-              pIdxKey->eqSeen = 1;
-              bestIdx = i;
-              bestCmp = pIdxKey->default_rc;
-              treeFound = 1;
-              treeCmp = bestCmp;
-              if( pIdxKey->default_rc < 0 ){
-                continue;
-              }
-              break;
-            }
-          }
-          prollyNodeValue(&pLeaf->node, i, &pVal, &nVal);
-          if( nVal==0 ){
-            rc = recordFromSortKeyBufferColl(
-                pSK, nSK, pCur->pKeyInfo, &pRecBuf, &nRecBufAlloc, &nVal);
-            if( rc!=SQLITE_OK ) break;
-            pVal = pRecBuf;
-          }
-          pIdxKey->eqSeen = 0;
-          recCmp = sqlite3VdbeRecordCompare(nVal, pVal, pIdxKey);
-
-          if( recCmp==0 || pIdxKey->eqSeen ){
-            if( pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap) ){
-              ProllyMutMapEntry *mmE = 0;
-              rc = prollyMutMapFindRc(pCur->pMutMap, pSK, nSK, 0, &mmE);
-              if( rc!=SQLITE_OK ) break;
-              if( mmE && mmE->op==PROLLY_EDIT_DELETE ){
-                continue;
-              }
-            }
-            bestIdx = i;
-            bestCmp = recCmp;
-            treeFound = 1;
-            treeCmp = recCmp;
-            break;
-          } else if( recCmp > 0 ){
-            if( pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap) ){
-              ProllyMutMapEntry *mmE = 0;
-              rc = prollyMutMapFindRc(pCur->pMutMap, pSK, nSK, 0, &mmE);
-              if( rc!=SQLITE_OK ) break;
-              if( mmE && mmE->op==PROLLY_EDIT_DELETE ){
-                continue;
-              }
-            }
-            if( bestIdx < 0 ){
-              bestIdx = i;
-              bestCmp = recCmp;
-            }
-          }
-        }
-        if( rc!=SQLITE_OK || treeFound || bestIdx>=0 ) break;
-        if( !(pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap)) ) break;
-        /* Everything from the seek point through the end of this leaf was
-        ** delete-masked; the first live row may sit in a later leaf. */
-        if( nItems>0 ) pCur->pCur.aLevel[pCur->pCur.iLevel].idx = nItems-1;
-        rc = prollyCursorNext(&pCur->pCur);
-        if( rc!=SQLITE_OK ) break;
-        if( pCur->pCur.eState!=PROLLY_CURSOR_VALID ) break;
-        iLevel = pCur->pCur.iLevel;
-        pLeaf = pCur->pCur.aLevel[iLevel].pEntry;
-        nItems = pLeaf->node.nItems;
-        lo = 0;
-       }
-        pCur->pMovetoRec = pRecBuf;
-        pCur->nMovetoRecAlloc = nRecBufAlloc;
-        if( rc!=SQLITE_OK ) return rc;
-      }
-
-      if( treeFound ){
-
-        pCur->pCur.aLevel[iLevel].idx = bestIdx;
-      } else if( bestIdx >= 0 ){
-
-        pCur->pCur.aLevel[iLevel].idx = bestIdx;
-        treeCmp = bestCmp;
-        treeFound = 1;
-      }
-
-    }
-
-    {
-      struct TableEntry *pTE = findTable(pCur->pBtree, pCur->pgnoRoot);
-      ProllyMutMap *pPending = pTE ? (ProllyMutMap*)pTE->pPending : 0;
-      if( ((pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap))
-         || (pPending && pPending!=pCur->pMutMap
-             && !prollyMutMapIsEmpty(pPending)))
-       && !(treeFound && treeCmp==0) ){
-      int savedEqSeen = pIdxKey->eqSeen;
-      rc = findMatchingMutMapEntry((ProllyMutMap*)pCur->pMutMap,
-                                   pCur->pKeyInfo,
-                                   pIdxKey, pSortKey, nSortKey,
-                                   exactMutMapKey,
-                                   &mutE, &mutCmp);
-      if( rc!=SQLITE_OK ){
-        return rc;
-      }
-      if( mutE ) mutFromCursorMap = 1;
-      if( !mutE && pPending && pPending!=pCur->pMutMap ){
-        rc = findMatchingMutMapEntry(pPending,
-                                     pCur->pKeyInfo,
-                                     pIdxKey, pSortKey, nSortKey,
-                                     exactMutMapKey,
-                                     &mutE, &mutCmp);
-        if( rc!=SQLITE_OK ){
-          return rc;
-        }
-      }
-      if( mutE ){
-
-        const u8 *pMutVal = mutE->pVal;
-        int nMutVal = mutE->nVal;
-        if( mutFromCursorMap && nMutVal==0 ){
-          mutFound = 1;
-        }else if( nMutVal==0 ){
-          rc = recordFromSortKeyBufferColl(
-              mutE->pKey, mutE->nKey, pCur->pKeyInfo,
-              &pCur->pMovetoRec, &pCur->nMovetoRecAlloc, &nMutVal);
-          if( rc!=SQLITE_OK ) return rc;
-          pMutVal = pCur->pMovetoRec;
-        }
-        if( pMutVal ){
-          mutKey = pMutVal;
-          mutNKey = nMutVal;
-          mutFound = 1;
-        }
-      }
-      pIdxKey->eqSeen = savedEqSeen;
-    }
-    }
-
-    if( mutFound ){
-      int useMut = !treeFound;
-      if( treeFound ){
-        const u8 *pTreeKey = 0;
-        int nTreeKey = 0;
-        int nCmp;
-        int cmp;
-        prollyCursorKey(&pCur->pCur, &pTreeKey, &nTreeKey);
-        nCmp = nTreeKey < mutE->nKey ? nTreeKey : mutE->nKey;
-        cmp = memcmp(pTreeKey, mutE->pKey, nCmp);
-        if( cmp>0 || (cmp==0 && nTreeKey>mutE->nKey) ){
-          useMut = 1;
-        }
-      }
-      if( useMut ){
-        if( mutFromCursorMap ){
-          setCursorToMutMapEntryPhys(
-              pCur, (int)(mutE - pCur->pMutMap->aEntries));
-          pCur->deferredTreeSeek = 1;
-        }else{
-          rc = cacheCursorPayloadCopy(pCur, mutKey, mutNKey);
-          if( rc!=SQLITE_OK ){
-            return rc;
-          }
-          pCur->eState = CURSOR_VALID;
-        }
-        *pRes = mutCmp;
-        pIdxKey->eqSeen = 1;
-        return SQLITE_OK;
-      }
-    }
-    if( treeFound ){
-      *pRes = treeCmp;
-      /* A prefix seek can land on a tree row whose value was overwritten in
-      ** this transaction (full-key seeks catch this in the exact-match fast
-      ** path above). Serve the row from the mut map, or the caller reads the
-      ** stale tree payload. */
-      if( pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap) ){
-        const u8 *pTreeKey = 0;
-        int nTreeKey = 0;
-        ProllyMutMapEntry *mmE = 0;
-        prollyCursorKey(&pCur->pCur, &pTreeKey, &nTreeKey);
-        rc = prollyMutMapFindRc(pCur->pMutMap, pTreeKey, nTreeKey, 0, &mmE);
-        if( rc!=SQLITE_OK ) return rc;
-        if( mmE && mmE->op==PROLLY_EDIT_INSERT ){
-          setCursorToMutMapEntryPhys(
-              pCur, (int)(mmE - pCur->pMutMap->aEntries));
-          pCur->deferredTreeSeek = 1;
-          return SQLITE_OK;
-        }
-      }
-      pCur->eState = CURSOR_VALID;
-      cacheCurrentTreeStoredPayloadNonIntKey(pCur);
-      return SQLITE_OK;
-    }
-  }
-
-  {
-    int lastRes = 0;
-    rc = prollyCursorLast(&pCur->pCur, &lastRes);
-    if( rc!=SQLITE_OK ) return rc;
-    if( pCur->pMutMap && !prollyMutMapIsEmpty(pCur->pMutMap) ){
-      /* No live row at or above the seek key on either side. Land on the
-      ** merged last row so delete-masked tree rows are never exposed and
-      ** mut-map-only rows are still reachable. */
-      pCur->mmActive = 1;
-      rc = mergeLast(pCur, &lastRes);
-      if( rc!=SQLITE_OK ) return rc;
-    }
-    if( lastRes==0 ){
-      pCur->eState = CURSOR_VALID;
-      *pRes = -1;
-      if( !pCur->mmActive || pCur->mergeSrc!=MERGE_SRC_MUT ){
-        cacheCurrentTreeStoredPayloadNonIntKey(pCur);
-      }
-    } else {
-      pCur->eState = CURSOR_INVALID;
-      *pRes = -1;
-    }
-  }
-  return SQLITE_OK;
-}
-int sqlite3BtreeIndexMoveto(
-  BtCursor *pCur,
-  UnpackedRecord *pIdxKey,
-  int *pRes
-){
-  if( !pCur ) return SQLITE_OK;
-  return pCur->pCurOps->xIndexMoveto(pCur, pIdxKey, pRes);
-}
-
-int cachedSeekKeyMatchesCurrent(BtCursor *pCur){
-  const u8 *pKey = 0;
-  int nKey = 0;
-
-  if( !pCur || pCur->curIntKey
-   || pCur->nSeekSortKey<=0 || pCur->nSeekKeyField!=0 ){
-    return 0;
-  }
-  if( pCur->mmActive
-   && (pCur->mergeSrc==MERGE_SRC_MUT || pCur->mergeSrc==MERGE_SRC_BOTH) ){
-    ProllyMutMapEntry *e = currentMutMapEntry(pCur);
-    if( !e ) return 0;
-    pKey = e->pKey;
-    nKey = e->nKey;
-  }else if( prollyCursorIsValid(&pCur->pCur) ){
-    prollyCursorKey(&pCur->pCur, &pKey, &nKey);
-  }
-  return pKey && nKey==pCur->nSeekSortKey
-      && memcmp(pKey, pCur->pSeekSortKey, nKey)==0;
-}
-
-int sqlite3BtreeProllyCachedIndexKeyCompare(
-  BtCursor *pCur,
-  UnpackedRecord *pIdxKey,
-  int *pRes
-){
-  const u8 *pKey = 0;
-  int nKey = 0;
-  int nCmp;
-  int cmp;
-
-  if( !pCur || pCur->pCurOps!=&prollyCursorOps || pCur->curIntKey ){
-    return SQLITE_NOTFOUND;
-  }
-  if( pCur->eState!=CURSOR_VALID || !pIdxKey || !pCur->pKeyInfo ){
-    return SQLITE_NOTFOUND;
-  }
-
-  if( pCur->mmActive
-   && (pCur->mergeSrc==MERGE_SRC_MUT || pCur->mergeSrc==MERGE_SRC_BOTH) ){
-    ProllyMutMapEntry *e = currentMutMapEntry(pCur);
-    if( !e ) return SQLITE_NOTFOUND;
-    pKey = e->pKey;
-    nKey = e->nKey;
-  }else if( prollyCursorIsValid(&pCur->pCur) ){
-    prollyCursorKey(&pCur->pCur, &pKey, &nKey);
-  }else{
-    return SQLITE_NOTFOUND;
-  }
-
-  if( prollyBtCursorCursorHasHint(pCur, BTREE_SEEK_EQ)
-   && pCur->nSeekSortKey>0
-   && pCur->nSeekKeyField==(int)pIdxKey->nField ){
-    nCmp = nKey < pCur->nSeekSortKey ? nKey : pCur->nSeekSortKey;
-    cmp = memcmp(pKey, pCur->pSeekSortKey, nCmp);
-    if( cmp<0 ){
-      *pRes = -1;
-    }else if( cmp>0 ){
-      *pRes = 1;
-    }else if( nKey < pCur->nSeekSortKey ){
-      *pRes = -1;
-    }else{
-      pIdxKey->eqSeen = 1;
-      *pRes = pIdxKey->default_rc;
-    }
-    return SQLITE_OK;
-  }
-
-  {
-    int rc;
-    if( pCur->nCompareSortKey<=0
-     || pCur->nCompareKeyField!=(int)pIdxKey->nField ){
-      rc = sortKeyFromUnpackedForCount(
-          pCur, pIdxKey, &pCur->pCompareSortKey,
-          &pCur->nCompareSortKeyAlloc, &pCur->nCompareSortKey);
-      if( rc!=SQLITE_OK ){
-        CLEAR_CACHED_COMPARE_KEY(pCur);
-        return rc;
-      }
-      pCur->nCompareKeyField = (int)pIdxKey->nField;
-    }
-    nCmp = nKey < pCur->nCompareSortKey ? nKey : pCur->nCompareSortKey;
-    cmp = memcmp(pKey, pCur->pCompareSortKey, nCmp);
-    if( cmp<0 ){
-      *pRes = -1;
-    }else if( cmp>0 ){
-      *pRes = 1;
-    }else if( nKey < pCur->nCompareSortKey ){
-      *pRes = -1;
-    }else{
-      pIdxKey->eqSeen = 1;
-      *pRes = pIdxKey->default_rc;
-    }
-    return SQLITE_OK;
-  }
-}
-
-/* Clear cached compare key before OP_SeekScan changes seek targets. */
-void sqlite3BtreeProllyClearCompareKey(BtCursor *pCur){
-  if( pCur ){
-    CLEAR_CACHED_COMPARE_KEY(pCur);
-  }
-}
 
 int sqlite3BtreeProllyIndexRowid(BtCursor *pCur, i64 *pRowid){
   const u8 *pKey = 0;
