+++ src/prolly_btree_int.h
@@ -478,6 +478,8 @@ ProllyMutMapEntry *currentMutMapEntry(BtCursor *pCur);
 void setCursorToMutMapEntryPhys(BtCursor *pCur, int physIdx);
 const char *findTableNumberName(sqlite3 *db, Pgno iTable);
 int cachedSeekKeyMatchesCurrent(BtCursor *pCur);
+void cacheCurrentTreePayloadIfIntKey(BtCursor *pCur);
+void cacheCurrentTreeStoredPayloadNonIntKey(BtCursor *pCur);
 int copyZeroTailPayload(ProllyMutMapEntry *e, u32 offset, u32 amt, void *pBuf);
 void btreeMarkWorkingStateChanged(Btree *p, int bLocal);
 int restoreFromCommitted(Btree *p);
@@ -729,6 +731,7 @@ int sortKeyFromUnpackedForCount(BtCursor*, UnpackedRecord*, u8**, int*, int*);
 int materializeDeferredMergedSeekBackward(BtCursor *pCur);
 int tableEntryIsTableRoot(Btree*, struct TableEntry*, int*);
 void clearMergeCursorState(BtCursor*);
+int mergeLast(BtCursor *pCur, int *pRes);
 int prollyCursorCheckInterrupt(BtCursor*);
 int advanceTreeCursor(BtCursor*, int);
 ProllyMutMapEntry *orderedMutMapEntryAt(ProllyMutMap*, int);
